+++ resources/views/components/input/pikaday.blade.php
@@ -3,7 +3,7 @@
 @php
     $for = $attributes->whereStartsWith('wire:model')->first() ?? $attributes->get('for');
     if (!isset($for)) {
-        throw new InvalidArgumentException('This form control must have at least one of the following attributes specified: `wire:model` or `for`. ');
+        throw new InvalidArgumentException('The form control requires either a `for` or `wire:model` attribute to be specified.');
     }
 @endphp